Started to move Ballerup-specific code away from apos-integration
