+++ integrations/ballerup/ballerup.sh
@@ -1,0 +1,8 @@
+export MUNICIPALITY_NAME='Ballerup'
+export ANSAT_UUID=56e1214a-330f-4592-89f3-ae3ee8d5b2e6
+export GLOBAL_DATE=2010-01-01
+export BASE_APOS_URL=http://apos.balk.dk:8080/apos2-
+export CREATE_UDVALGS_CLASSES=yes
+export PHNONE_NAMES="Alt. Tlf.":"Kontakt Tlf."
+export EMAIL_NAME="E-mail"
+export MAIN_PHONE_NAME="Kontakt Tlf."
